improve the logic of methods related to UserService
