+++ src/main/java/de/tum/in/www1/artemis/web/rest/CourseResource.java
@@ -379,7 +379,7 @@ public class CourseResource {
                     HeaderUtil.createFailureAlert(applicationName, false, ENTITY_NAME, "registrationDisabled", "The course does not allow registration. Cannot register user"))
                     .body(null);
         }
-        artemisAuthenticationProvider.registerUserForCourse(user, course);
+        courseService.registerUserForCourse(user, course);
         return ResponseEntity.ok(user);
     }
