+++ src/test/java/de/tum/in/www1/artemis/connector/gitlab/GitlabRequestMockProvider.java
@@ -1,5 +1,7 @@
 package de.tum.in.www1.artemis.connector.gitlab;
 
+import static org.gitlab4j.api.models.AccessLevel.GUEST;
+import static org.gitlab4j.api.models.AccessLevel.MAINTAINER;
 import static org.mockito.Mockito.*;
 
 import java.net.URL;
@@ -64,7 +66,8 @@ public class GitlabRequestMockProvider {
         final Group group = new Group().withPath(exercisePath).withName(exerciseName).withVisibility(Visibility.PRIVATE);
         doReturn(group).when(groupApi).addGroup(group);
         mockGetUserID();
-        mockAddUserToGroup(exercise);
+        mockAddUserToGroup(exercise.getProjectKey(), MAINTAINER, "instructor1", 2);
+        mockAddUserToGroup(exercise.getProjectKey(), GUEST, "tutor1", 3);
     }
 
     /**
@@ -88,17 +91,12 @@ public class GitlabRequestMockProvider {
         doReturn(new org.gitlab4j.api.models.User()).when(userApi).updateUser(any(), any());
     }
 
-    private void mockAddUserToGroup(ProgrammingExercise exercise) throws GitLabApiException {
-        final Member instructor = new Member();
-        instructor.setAccessLevel(AccessLevel.MAINTAINER);
-        instructor.setId(2);
-        instructor.setName("instructor1");
-        final Member tutor = new Member();
-        tutor.setAccessLevel(AccessLevel.GUEST);
-        tutor.setId(3);
-        tutor.setName("tutor1");
-        doReturn(instructor).when(groupApi).addMember(exercise.getProjectKey(), 2, 40);
-        doReturn(tutor).when(groupApi).addMember(exercise.getProjectKey(), 3, 10);
+    public void mockAddUserToGroup(String group, AccessLevel accessLevel, String userName, Integer userId) throws GitLabApiException {
+        final Member member = new Member();
+        member.setAccessLevel(accessLevel);
+        member.setId(userId);
+        member.setName(userName);
+        doReturn(member).when(groupApi).addMember(group, userId, accessLevel);
     }
 
     public void mockCreateRepository(ProgrammingExercise exercise, String repositoryName) throws GitLabApiException {
